feat: oauth signup plus bunch of improvements

+++ backend/src/main/java/io/papermc/hangar/components/auth/service/VerificationService.java
@@ -66,6 +66,9 @@ public class VerificationService extends HangarComponent {
         if (userTable == null) {
             return;
         }
+        if (!userTable.isEmailVerified()) {
+            return;
+        }
 
         this.verificationCodeDao.deleteOld(VerificationCodeTable.VerificationCodeType.PASSWORD_RESET, userTable.getUserId());
